LayoutTests/accessibility/native-listbox-no-selection-expected.txt


Tests that native listboxes with no selected children report no selected children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listbox.selectedChildrenCount is 0
PASS listbox.selectedChildAtIndex(1) == null is true
PASS successfullyParsed is true

TEST COMPLETE
